CI: Configure sonarcloud on github action (#1065)
* Sets up sonarcloud on github actions to run static code analysis for python and C++
* Collect coverage information for sonarcloud
* Moves time-consuming valgrind run to separate workflow
* Cleans up installation scripts
* Fixes a number of minor inconveniences

Progress towards #1053. Still keeping everything on travis ci, while testing whether we are happy with sonarcloud and github actions. To be slimmed down once accepted.

Note that a proper report will only be displayed after sonarscanner has been run on the base branch.
