+++ .github/workflows/test_cplusplus_valgrind.yml
@@ -1,0 +1,54 @@
+name: C++ testsuite / Valgrind / Ubuntu
+on:
+  push:
+    branches:
+      - develop
+      - master
+      - feature_1053_sonarcloud
+
+
+jobs:
+  build:
+    name: _
+
+    # TODO: prepare image with more deps preinstalled
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@master
+    - run: git fetch --prune --unshallow
+
+    # install amici dependencies
+    - name: apt
+      run: |
+        sudo apt-get update \
+          && sudo apt-get install -y \
+            cmake \
+            g++ \
+            libatlas-base-dev \
+            libboost-serialization-dev \
+            libhdf5-serial-dev \
+            swig \
+            valgrind
+
+    - name: Build suitesparse
+      run: |
+        scripts/buildSuiteSparse.sh
+
+    - name: Build sundials
+      run: |
+        scripts/buildSundials.sh
+
+    - name: Build cpputest
+      run: |
+        scripts/buildCpputest.sh
+
+    - name: Build AMICI
+      # TODO: should get rid of having to install numpy before
+      run: |
+        pip3 install numpy \
+        && scripts/buildAmici.sh
+
+    - name: C++ tests / Valgrind
+      run: |
+        scripts/run-valgrind.sh
