+++ python/sdist/setup.py
@@ -12,20 +12,29 @@ Requires:
 - hdf5 libraries and headers
 """
 
-from setuptools import find_packages, setup, Extension
-
 import os
-import sys
 import subprocess
+import sys
+
 import setup_clibs  # Must run from within containing directory
 
+from distutils import log
 
-def try_install(package):
+
+def try_install(package: str) -> None:
     """Try installing the given package using pip. Exit on error."""
+
+    log.info(f"Missing required package {package}. Trying to install...")
     errno = subprocess.call([sys.executable, "-m", "pip", "install", package])
     if errno:
-        print(f"Failed trying to install {package}. Please install manually.")
+        log.error(f"Failed trying to install {package}. "
+                  "Please install manually before installing AMICI.")
         raise SystemExit(errno)
+
+    import importlib
+    importlib.invalidate_caches()
+    globals()[package] = importlib.import_module(package)
+
 
 try:
     # Required for numpy include directory, and for importing anything from
@@ -36,6 +45,14 @@ except ImportError:
     try_install('numpy')
     # retry
     import numpy as np
+
+try:
+    # Required for customizing installation, but setup_requires fires too late
+    from setuptools import find_packages, setup, Extension
+except ImportError:
+    try_install('setuptools')
+    from setuptools import find_packages, setup, Extension
+
 
 # Add current directory to path, as we need some modules from the AMICI
 # package already for installation
@@ -51,7 +68,6 @@ from amici.setuptools import (
     add_coverage_flags_if_required,
     add_debug_flags_if_required,
 )
-
 
 
 # Python version check. We need >= 3.6 due to e.g. f-strings
