+++ python/tests/test_misc.py
@@ -80,4 +80,6 @@ def test_cmake_compilation(sbml_example_presimulation_module):
 
     cmd = f"set -e; cd {source_dir}; mkdir -p build; cd build; "\
           "cmake ..; make"
-    subprocess.run(cmd, shell=True, capture_output=True, check=True)
+
+    subprocess.run(cmd, shell=True, check=True,
+                   stdout=subprocess.PIPE, stderr=subprocess.PIPE)
