+++ sonar-project.properties
@@ -1,0 +1,21 @@
+# SonarScanner configuration
+# https://docs.sonarqube.org/latest/analysis/scan/sonarscanner/
+# https://docs.sonarqube.org/latest/analysis/analysis-parameters/
+# https://sonarcloud.io/documentation/analysis/languages/cfamily/
+
+sonar.host.url=https://sonarcloud.io
+sonar.login=af35cb17710485d21c8e453a77f1f008eae1f7a4
+sonar.organization=icb-dcm
+sonar.projectKey=ICB-DCM_AMICI
+
+sonar.sources=.
+sonar.exclusions=build/**,ThirdParty/**,doc/**,models/**,python/sdist/amici/**
+
+sonar.sourceEncoding=UTF-8
+
+sonar.cfamily.threads=2
+sonar.cfamily.gcov.reportsPath=build
+sonar.cfamily.cache.enabled=true
+sonar.cfamily.cache.path=sonar_cache
+
+sonar.python.coverage.reportPaths=build/coverage_py.xml
